move each ResourceDescription message to its own module
